requirements.txt
black==22.10.0
click==8.1.3
colorama==0.4.6
contourpy==1.0.6
cycler==0.11.0
fonttools==4.38.0
keyboard==0.13.5
kiwisolver==1.4.4
matplotlib==3.6.2
mypy-extensions==0.4.3
numpy==1.23.4
packaging==21.3
pandas==1.5.1
pathspec==0.10.2
Pillow==9.3.0
platformdirs==2.5.4
pyparsing==3.0.9
pyserial==3.5
python-dateutil==2.8.2
pytz==2022.6
six==1.16.0
tomli==2.0.1